More cleanly integrates feature spec helpers into rspec.
