+++ spec/support/helpers/feature_spec_helpers.rb
@@ -1,0 +1,3 @@
+module FeatureSpecHelpers
+  include UtilityEnhancedComponentCollectorHelper
+end
