+++ spec/test_app/app/helpers/component_attribute_defaults_helper.rb
@@ -1,0 +1,17 @@
+module ComponentAttributeDefaultsHelper
+  def modal_id
+    'test_modal'
+  end
+
+  def modal_target
+    "##{modal_id}"
+  end
+
+  def is_true?
+    false
+  end
+
+  def href
+    '#included_href'
+  end
+end
