+++ spec/test_app/app/helpers/feature_spec_helpers.rb
@@ -1,0 +1,6 @@
+# module FeatureSpecHelpers
+#   include FeatureSpecHelpers::ComponentRespondsHelper
+#   include FeatureSpecHelpers::ComponentAttributeDefaultsHelper
+#   # require 'feature_spec_helpers/component_attribute_defaults_helper'
+#   # require 'feature_spec_helpers/component_responds_helper'
+# end
